requirements.txt
flask==3.0.0
pandas==2.1.0
folium==0.14.0
requests==2.31.0
google-generativeai==0.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0
pytest==7.4.0 